pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yandex-query-magic"
version = "0.0.1"
authors = [
  { email="cloud@support.yandex.ru" },
]
description = "The Yandex Query official Jupyter Notebook plugin"
readme = "README.md"
license="MIT"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  "ipywidgets>=8.1.1",
  "pyjwt>=2.3.0",
  "requests>=2.31.0",
  "asyncio>=3.4.3",
  "aiohttp>=3.9.1",
  "pandas>=2.2.0",
  "nest_asyncio>=1.6.0",
  "typing>=3.7.4.3",
  "aiohttp_retry>=2.8.3",
  "python-dateutil>=2.8.2",
  "urllib3>=1.26.5",
  "jupyter-ui-poll>=0.2.2",
  "cryptography>=3.4.8",
  "jupyter-contrib-nbextensions>=0.7.0"
]
[project.optional-dependencies]
test = [
    "Werkzeug>=3.0.1",
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.4",
    "pytest-httpserver>=1.0.8",
    "Jinja2"
]

[project.urls]
Homepage = "https://github.com/SloNN/yandex_query_magic"
Issues = "https://github.com/SloNN/yandex_query_magic/issues"

